Idk not work

+++ mc-packet-macros/src/lib.rs
@@ -7,7 +7,7 @@ use syn::{ItemStruct, LitInt, parse_macro_input};
 #[proc_macro_attribute]
 pub fn mc_packet(args: TokenStream, input: TokenStream) -> TokenStream {
     let args = parse_macro_input!(args as LitInt);
-    let packet_id: usize = args.base10_parse().expect("Arg must be a valid usize");
+    let packet_id: i32 = args.base10_parse().expect("Arg must be a valid usize");
 
     let input_ast = parse_macro_input!(input as ItemStruct);
     let struct_name = input_ast.ident.clone();
